--- a/benchmarks/bench_fp.nim
+++ b/benchmarks/bench_fp.nim
@@ -55,5 +55,6 @@ proc main() =
 main()
 
 echo "Notes:"
-echo "  GCC is significantly slower than Clang on multiprecision arithmetic."
-echo "  The simplest operations might be optimized away by the compiler."
+echo "  - GCC is significantly slower than Clang on multiprecision arithmetic."
+echo "  - The simplest operations might be optimized away by the compiler."
+echo "  - Fast Squaring and Fast Multiplication are possible if there are spare bits in the prime representation (i.e. the prime uses 254 bits out of 256 bits)"