--- a/constantine/towers.nim
+++ b/constantine/towers.nim
@@ -69,6 +69,12 @@ func `*`*(_: typedesc[β], a: Fp2): Fp2 {.inline, noInit.} =
   #
   # (c0 + c1 x) (u + v x) => u c0 + (u c0 + u c1)x + v c1 x²
   #                       => u c0 + β v c1 + (v c0 + u c1) x
+
+  # TODO: check code generated when ξ = 1 + 𝑖
+  #       The mul by constant are inline but
+  #       since we don't have __restrict tag
+  #       and we use arrays (which decay into pointer)
+  #       the compiler might not elide the temporary
   when a.fromComplexExtension():
     result.c0.diff(u * a.c0, v * a.c1)
   else:
@@ -78,7 +84,21 @@ func `*`*(_: typedesc[β], a: Fp2): Fp2 {.inline, noInit.} =
 func `*=`*(a: var Fp2, _: typedesc[ξ]) {.inline.} =
   ## Multiply an element of 𝔽p by the quadratic non-residue
   ## chosen to construct 𝔽p2
-  a = ξ * a
+  # Yet another const tuple unpacking bug
+  const u = Fp2.C.get_CNR_Fp2()[0] # Cubic non-residue to construct 𝔽p6
+  const v = Fp2.C.get_CNR_Fp2()[1]
+  const Beta = Fp2.C.get_QNR_Fp()  # Quadratic non-residue to construct 𝔽p2
+  # ξ = u + v x
+  # and x² = β
+  #
+  # (c0 + c1 x) (u + v x) => u c0 + (u c0 + u c1)x + v c1 x²
+  #                       => u c0 + β v c1 + (v c0 + u c1) x
+  when a.fromComplexExtension() and u == 1 and v == 1:
+    let t = a.c0
+    a.c0 -= a.c1
+    a.c1 += t
+  else: # TODO: optim for inline
+    a = ξ * a
 
 # 𝔽p12
 # ----------------------------------------------------------------