This PR didn't manage to break the Nim compiler despite my best effort:
- Tower of co-recursive concepts: the compilation time are slow but tolerable, well 12s from cold cache, 9 from hot. Probably10x faster than C++ anyway.
![image](https://user-images.githubusercontent.com/22738317/79168073-495e7300-7de9-11ea-8e23-95d245274428.png)
- Now the tower can easily be used for KSS18, BLS24 and BLS48 families
- Unicode type names passed as typedesc, welcome β, ξ, γ

The curves that didn't have their non-residue parametrized have been commented out.
Probably the curves that are not popular or standardized or in the process of should be deleted.