@mratsim 2020-04-13
Perf difference on BLS12-381 curve since the tower was correct for this curve previously:

## Old towering impl

```
--------------------------------------------------------------------------------
Addition        Fp2[BLS12_381]          9 ns        29 cycles
Substraction    Fp2[BLS12_381]          9 ns        29 cycles
Negation        Fp2[BLS12_381]          4 ns        14 cycles
Multiplication  Fp2[BLS12_381]        157 ns       473 cycles
Squaring        Fp2[BLS12_381]        107 ns       321 cycles
Inversion       Fp2[BLS12_381]      16235 ns     48704 cycles
--------------------------------------------------------------------------------

--------------------------------------------------------------------------------
Addition        Fp6[BLS12_381]         40 ns       121 cycles
Substraction    Fp6[BLS12_381]         40 ns       120 cycles
Negation        Fp6[BLS12_381]         18 ns        56 cycles
Multiplication  Fp6[BLS12_381]       1141 ns      3424 cycles
Squaring        Fp6[BLS12_381]        771 ns      2313 cycles
Inversion       Fp6[BLS12_381]      17676 ns     53029 cycles
--------------------------------------------------------------------------------

--------------------------------------------------------------------------------
Addition        Fp12[BLS12_381]        83 ns       249 cycles
Substraction    Fp12[BLS12_381]        80 ns       240 cycles
Negation        Fp12[BLS12_381]        48 ns       146 cycles
Multiplication  Fp12[BLS12_381]      3691 ns     11074 cycles
Squaring        Fp12[BLS12_381]      2540 ns      7620 cycles
Inversion       Fp12[BLS12_381]     21822 ns     65469 cycles
--------------------------------------------------------------------------------

```

## New

```
--------------------------------------------------------------------------------
Addition        Fp2[BLS12_381]          8 ns        25 cycles
Substraction    Fp2[BLS12_381]          8 ns        24 cycles
Negation        Fp2[BLS12_381]          3 ns         9 cycles
Multiplication  Fp2[BLS12_381]        160 ns       481 cycles
Squaring        Fp2[BLS12_381]        107 ns       323 cycles
Inversion       Fp2[BLS12_381]      15620 ns     46861 cycles
--------------------------------------------------------------------------------

--------------------------------------------------------------------------------
Addition        Fp6[BLS12_381]         26 ns        80 cycles
Substraction    Fp6[BLS12_381]         25 ns        75 cycles
Negation        Fp6[BLS12_381]          9 ns        27 cycles
Multiplication  Fp6[BLS12_381]       1181 ns      3545 cycles
Squaring        Fp6[BLS12_381]        790 ns      2370 cycles
Inversion       Fp6[BLS12_381]      17740 ns     53221 cycles
--------------------------------------------------------------------------------

--------------------------------------------------------------------------------
Addition        Fp12[BLS12_381]        51 ns       155 cycles
Substraction    Fp12[BLS12_381]        49 ns       148 cycles
Negation        Fp12[BLS12_381]        18 ns        56 cycles
Multiplication  Fp12[BLS12_381]      3809 ns     11427 cycles
Squaring        Fp12[BLS12_381]      2631 ns      7895 cycles
Inversion       Fp12[BLS12_381]     21802 ns     65409 cycles
--------------------------------------------------------------------------------
```

Not too sure why addition is so much faster, it probably has to do with inlining.

Multiplication is slightly slower probably due to useless work done when multiplying by the unitary non-residue that incur an extra useless copy: https://github.com/mratsim/constantine/blob/3945bffb068870baa79eb9d6e3adc02cb06b422a/constantine/towers.nim#L60-L76